add user level to profile
